requirements.txt
fpdf2>=2.7.4
requests>=2.31.0
PyYAML>=6.0
google-genai>=0.1.0
openai>=1.0.0
anthropic>=0.18.0
